Moved rest of components to shared
