+++ apps/firefox/src/popup/App.tsx
@@ -1,6 +1,6 @@
+import { Navbar } from "@ui/components/Navbar";
 import { Outlet } from "react-router-dom";
 import styled from "styled-components";
-import { Navbar } from "./components/Navbar";
 
 function App() {
   return (
